Add features to fix builds
