+++ web/src/lib.rs
@@ -1,5 +1,5 @@
 #![deny(unused_imports)]
-#![feature(allocator_api, new_uninit)]
+#![feature(new_uninit)]
 
 use maplibre::{
     event_loop::EventLoop,
